"shortcode-filter" Filter
The \Fgms\Distributor\Controller class now uses the WordPress builtin function apply_filters to obtain the [find_a_distributor] shortcode's HTML content.  It does this by invoking the "shortcode-filter" filter.  Note that since the \Fgms\Distributor\Controller class prefixes all strings used for configuration values et cetera with the prefix passed to its constructor the actual filter name is "{PREFIX}shortcode-filter".  If one does not provide a prefix to the constructor (and thereby accepts the default) the filter is "fgms-distributor-shortcode-filter".

A default implementation (which defers to any other present implementation) of this filter is provided in wordpress-find-a-distributor.php.