"distributor-filter" Filter
The \Fgms\Distributor\Controller class now uses apply_filters to obtain HTML content for each distributor returned over the AJAX API.  This HTML is stored as a string under the "html" property of the object for each distributor.  The filter is called "distributor-filter" or with the default prefix "fgms-distributor-distributor-filter".

A default implementation is provided in wordpress-find-a-distributor.php.  If any other filters are present this filter does nothing.